+++ .idea/workspace.xml
@@ -20,12 +20,13 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="1281c1a0-e059-4e36-bd84-f7862387484a" name="Default Changelist" comment="">
-      <change afterPath="$PROJECT_DIR$/file1." afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/file1.txt" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/file3." afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/file3.txt" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/file_duplicate_detector.py" beforeDir="false" afterPath="$PROJECT_DIR$/file_duplicate_detector.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/file1." beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/file1.txt" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/file3." beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/file3.txt" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/file_duplicate_detector.py" beforeDir="false" afterPath="$PROJECT_DIR$/FileManager.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/file_report_generator.py" beforeDir="false" afterPath="$PROJECT_DIR$/utils.py" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -60,6 +61,27 @@
     <property name="settings.editor.selected.configurable" value="Errors" />
   </component>
   <component name="RunManager">
+    <configuration name="FileManager" type="PythonConfigurationType" factoryName="Python" nameIsGenerated="true">
+      <module name="Duplicate-Control-Algorithm" />
+      <option name="INTERPRETER_OPTIONS" value="" />
+      <option name="PARENT_ENVS" value="true" />
+      <envs>
+        <env name="PYTHONUNBUFFERED" value="1" />
+      </envs>
+      <option name="SDK_HOME" value="" />
+      <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$" />
+      <option name="IS_MODULE_SDK" value="true" />
+      <option name="ADD_CONTENT_ROOTS" value="true" />
+      <option name="ADD_SOURCE_ROOTS" value="true" />
+      <option name="SCRIPT_NAME" value="$PROJECT_DIR$/FileManager.py" />
+      <option name="PARAMETERS" value="" />
+      <option name="SHOW_COMMAND_LINE" value="false" />
+      <option name="EMULATE_TERMINAL" value="false" />
+      <option name="MODULE_MODE" value="false" />
+      <option name="REDIRECT_INPUT" value="false" />
+      <option name="INPUT_FILE" value="" />
+      <method v="2" />
+    </configuration>
     <configuration default="true" type="PythonConfigurationType" factoryName="Python">
       <module name="Duplicate-Control-Algorithm" />
       <option name="INTERPRETER_OPTIONS" value="" />
@@ -70,27 +92,6 @@
       <option name="SDK_HOME" value="/usr/bin/python3.7" />
       <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$" />
       <option name="IS_MODULE_SDK" value="false" />
-      <option name="ADD_CONTENT_ROOTS" value="true" />
-      <option name="ADD_SOURCE_ROOTS" value="true" />
-      <option name="SCRIPT_NAME" value="$PROJECT_DIR$/file_duplicate_detector.py" />
-      <option name="PARAMETERS" value="" />
-      <option name="SHOW_COMMAND_LINE" value="false" />
-      <option name="EMULATE_TERMINAL" value="false" />
-      <option name="MODULE_MODE" value="false" />
-      <option name="REDIRECT_INPUT" value="false" />
-      <option name="INPUT_FILE" value="" />
-      <method v="2" />
-    </configuration>
-    <configuration name="file_duplicate_detector" type="PythonConfigurationType" factoryName="Python" nameIsGenerated="true">
-      <module name="Duplicate-Control-Algorithm" />
-      <option name="INTERPRETER_OPTIONS" value="" />
-      <option name="PARENT_ENVS" value="true" />
-      <envs>
-        <env name="PYTHONUNBUFFERED" value="1" />
-      </envs>
-      <option name="SDK_HOME" value="" />
-      <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$" />
-      <option name="IS_MODULE_SDK" value="true" />
       <option name="ADD_CONTENT_ROOTS" value="true" />
       <option name="ADD_SOURCE_ROOTS" value="true" />
       <option name="SCRIPT_NAME" value="$PROJECT_DIR$/file_duplicate_detector.py" />
@@ -140,22 +141,22 @@
       <screen x="72" y="27" width="1848" height="923" />
     </state>
     <state x="779" y="246" width="424" height="491" key="FileChooserDialogImpl/72.27.1848.923@72.27.1848.923" timestamp="1589642054148" />
-    <state width="1825" height="233" key="GridCell.Tab.0.bottom" timestamp="1589662886423">
+    <state width="1825" height="236" key="GridCell.Tab.0.bottom" timestamp="1589668697278">
       <screen x="72" y="27" width="1848" height="923" />
     </state>
-    <state width="1825" height="233" key="GridCell.Tab.0.bottom/72.27.1848.923@72.27.1848.923" timestamp="1589662886423" />
-    <state width="1825" height="233" key="GridCell.Tab.0.center" timestamp="1589662886422">
+    <state width="1825" height="236" key="GridCell.Tab.0.bottom/72.27.1848.923@72.27.1848.923" timestamp="1589668697278" />
+    <state width="1825" height="236" key="GridCell.Tab.0.center" timestamp="1589668697275">
       <screen x="72" y="27" width="1848" height="923" />
     </state>
-    <state width="1825" height="233" key="GridCell.Tab.0.center/72.27.1848.923@72.27.1848.923" timestamp="1589662886422" />
-    <state width="1825" height="233" key="GridCell.Tab.0.left" timestamp="1589662886422">
+    <state width="1825" height="236" key="GridCell.Tab.0.center/72.27.1848.923@72.27.1848.923" timestamp="1589668697275" />
+    <state width="1825" height="236" key="GridCell.Tab.0.left" timestamp="1589668697275">
       <screen x="72" y="27" width="1848" height="923" />
     </state>
-    <state width="1825" height="233" key="GridCell.Tab.0.left/72.27.1848.923@72.27.1848.923" timestamp="1589662886422" />
-    <state width="1825" height="233" key="GridCell.Tab.0.right" timestamp="1589662886422">
+    <state width="1825" height="236" key="GridCell.Tab.0.left/72.27.1848.923@72.27.1848.923" timestamp="1589668697275" />
+    <state width="1825" height="236" key="GridCell.Tab.0.right" timestamp="1589668697277">
       <screen x="72" y="27" width="1848" height="923" />
     </state>
-    <state width="1825" height="233" key="GridCell.Tab.0.right/72.27.1848.923@72.27.1848.923" timestamp="1589662886422" />
+    <state width="1825" height="236" key="GridCell.Tab.0.right/72.27.1848.923@72.27.1848.923" timestamp="1589668697277" />
     <state x="687" y="172" width="608" height="640" key="RollbackChangesDialog" timestamp="1589662552659">
       <screen x="72" y="27" width="1848" height="923" />
     </state>
