--- file3.
@@ -1,28 +1,0 @@
-ID:
-
-Company:
-
-Address:
-
-Category:
-
-Staff:
-
-Business:
-
-Seniority:
-
-References:
-
-Channels:
-
-Appointment:
-
-Interview:
-
-Tracer:
-
-Closing:
-
-Account:
-
